Generate bitplanes in double format
Generate bitplanes in double format

+++ finalcode/generateBitplanes.h
@@ -21,25 +21,25 @@ std::vector<cv::Mat> generateBitPlanes(cv::Mat &grayimage)
 
 	for (int i = 1; i<grayImageWithBorder.rows - 1; i++) 
 	{
-		uchar* orgp = grayImageWithBorder.ptr<uchar>(i);
-		uchar* newp0 = image_channels[0].ptr<uchar>(i - 1);
-		uchar* newp1 = image_channels[1].ptr<uchar>(i - 1);
-		uchar* newp2 = image_channels[2].ptr<uchar>(i - 1);
-		uchar* newp3 = image_channels[3].ptr<uchar>(i - 1);
-		uchar* newp4 = image_channels[4].ptr<uchar>(i - 1);
-		uchar* newp5 = image_channels[5].ptr<uchar>(i - 1);
-		uchar* newp6 = image_channels[6].ptr<uchar>(i - 1);
-		uchar* newp7 = image_channels[7].ptr<uchar>(i - 1);
+		double* orgp = grayImageWithBorder.ptr<double>(i);
+		double* newp0 = image_channels[0].ptr<double>(i - 1);
+		double* newp1 = image_channels[1].ptr<double>(i - 1);
+		double* newp2 = image_channels[2].ptr<double>(i - 1);
+		double* newp3 = image_channels[3].ptr<double>(i - 1);
+		double* newp4 = image_channels[4].ptr<double>(i - 1);
+		double* newp5 = image_channels[5].ptr<double>(i - 1);
+		double* newp6 = image_channels[6].ptr<double>(i - 1);
+		double* newp7 = image_channels[7].ptr<double>(i - 1);
 
 		for (int j = 1; j<grayImageWithBorder.cols - 1; j++)
 		{
-			uchar test = orgp[j];
+			double test = orgp[j];
 
 			newp0[j - 1] = (test > (orgp - 1)[j - 1]);
 			newp1[j - 1] = (test > (orgp - 1)[j]);
 			newp2[j - 1] = (test > (orgp - 1)[j + 1]);
 			newp3[j - 1] = (test > orgp[j - 1]);
-			newp4[j - 1] = (test > orgp[j + 1]); 
+			newp4[j - 1] = (test > orgp[j + 1]);
 			newp5[j - 1] = (test > (orgp + 1)[j - 1]);
 			newp6[j - 1] = (test > (orgp + 1)[j]);
 			newp7[j - 1] = (test > (orgp + 1)[j + 1]);
